requirements.txt
requests>=2.31
python-dotenv>=1.0
pyyaml>=6.0
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
matplotlib>=3.8
folium>=0.15
